fix auto deletion issue (#97)

+++ app/Module/Core/Cron/Attachments.php
@@ -5,7 +5,7 @@ namespace HelpGent\Module\Core\Cron;
 use HelpGent\Module\Core\Model\Attachment_Model;
 use HelpGent\Base\Helper;
 
-class Attachment {
+class Attachments {
 
     /**
      * Constuctor
@@ -13,43 +13,36 @@ class Attachment {
      * @return void
      */
     public function __construct() {
-		add_action( 'helpgent_hourly_cron', [ $this, 'manage_attachments' ] );
+		add_action( 'helpgent_daily_cron', [ $this, 'manage' ] );
 
     }
-
-
-
-	/**
-	 * Clean/Delete expired attachments
-	 *
-	 * @return void
-	 */
-
-	public function manage_attachments() {
-		$this->delete_attachments();
-
-	}
 
 	/**
 	 * Delete all the expired attachments
 	 *
 	 * @return void
 	 */
-    public function delete_attachments() {
+    public function manage() {
 
-		$auto_delete_after = Helper\get_option( 'hgAttatchmentDeletionAfter', 45 );
-		$date 	= "-$auto_delete_after days";
+		$auto_delete_after = (int) Helper\get_option( 'attatchmentDeletionAfter' );
+
+		if( $auto_delete_after <= 0 ) {
+			return;
+		}
+
+		$date 	= "-{$auto_delete_after} days";
 		$value 	= "'" . date('Y-m-d H:i:s', strtotime( $date ) ) . "'";
 
 		$args['where'] = [
-			'created_on' => [
-				'field' 	=> 'created_on',
+			'created_at' => [
+				'field' 	=> 'created_at',
 				'value'		=> $value,
 				'compare'	=> '<',
 			],
 		];
 
 		$expired = Attachment_Model::get_items( $args );
+		
 		if( ! empty( $expired ) ) {
 			foreach( $expired as $attachment ) {
 				Attachment_Model::delete_item( $attachment['id'] );
